fix(api): normalize observations endpoint timestamp type to float8
Cast EXTRACT(EPOCH FROM o.time) to float8 before multiplication to ensure
numeric epoch millisecond timestamps in API responses. Update OpenAPI and
API reference documentation to reflect numeric time field. Add focused
test assertions for type correctness in service and integration layers.

+++ server/src/services/observationService.ts
@@ -32,7 +32,7 @@ export async function getObservationsByBSSID(
   const { rows } = await query(
     `SELECT ROW_NUMBER() OVER (ORDER BY o.time) as id, o.bssid,
             COALESCE(NULLIF(o.ssid, ''), '(hidden)') as ssid, o.radio_type as type,
-            o.lat, o.lon, o.level as signal, EXTRACT(EPOCH FROM o.time)::BIGINT * 1000 as time,
+            o.lat, o.lon, o.level as signal, (EXTRACT(EPOCH FROM o.time) * 1000)::float8 as time,
             COALESCE(o.accuracy, 3.79) as acc, o.altitude as alt,
             gc.address as geocoded_address, gc.city as geocoded_city, gc.state as geocoded_state,
             gc.poi_name as geocoded_poi_name,
